+++ flutter/android/src/main/kotlin/io/sentry/flutter/SentryFlutterPlugin.kt
@@ -97,6 +97,7 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler {
       }
       args.getIfNotNull<Boolean>("anrEnabled") { options.isAnrEnabled = it }
       args.getIfNotNull<Boolean>("sendDefaultPii") { options.isSendDefaultPii = it }
+      args.getIfNotNull<Boolean>("enableNdkScopeSync") { options.isEnableScopeSync = it }
 
       val nativeCrashHandling = (args["enableNativeCrashHandling"] as? Boolean) ?: true
       // nativeCrashHandling has priority over anrEnabled