+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Feat: Auto Transactions (#643)
+* Feat: Add enableNdkScopeSync Android support (#665)
+
 # 6.2.1
 
 * Fix: `sentry_logging` works now on web (#660)