+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Fix: ConcurrentModificationError in when finishing span (#664)
+* Feat: Add enableNdkScopeSync Android support (#665)
 
 # 6.2.1