+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 6.2.2
 
 * Fix: ConcurrentModificationError in when finishing span (#664)
 * Feat: Add enableNdkScopeSync Android support (#665)